fix: better folder structure
