--- src/components/AboutChileChico/index.ts
@@ -1,1 +1,0 @@
-export { default as AboutChileChico } from "./AboutChileChico"
