--- src/components/AboutTheWedding/index.ts
@@ -1,1 +1,0 @@
-export { default as AboutTheWedding } from "./AboutTheWedding";
