+++ src/components/AboutTheWedding/index.tsx
