--- src/components/Footer/index.ts
@@ -1,1 +1,0 @@
-export { default as Footer } from "./Footer"
