--- src/components/Hero/index.ts
@@ -1,1 +1,0 @@
-export { default as Hero } from "./Hero";
