--- src/components/HowToGetThere/index.ts
@@ -1,1 +1,0 @@
-export { default as HowToGetThere } from "./HowToGetThere";
