+++ src/components/RSVP/index.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+interface Props {}
+
+const RSVP = (props: Props) => {
+  console.log("hola");
+  return <div>Hola</div>;
+};
+
+export default RSVP;
