--- src/components/WhereToStay/index.ts
@@ -1,1 +1,0 @@
-export { default as WhereToStay } from "./WhereToStay";
