+++ src/components/WhereToStay/index.tsx
